backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
cryptography
httpx
aiofiles
Pillow
python-dotenv
apscheduler
redis
pydantic
pydantic-settings
instagrapi
requests
